Initial commit for openapi2jsonschema tool
The code for this is fairly blunt at the moment but the interface
hopefully makes sense, and the generated schemas should work.

This has not been widely tested against a range of schemas yet, if it
hits an error it's likely to explode. Shipping because it's useful, and
it's nothing that a solid bit of writing unit tests and refactoring
can't fix.