Hide Kubernetes processing behind flag
It would be nice to have this pluggable, but for the moment lets just
hide the Kubernetes processing behind a flag.

This commit also adds a new Kubernetes fix, for arrays which in
Kubernetes can be null as well as being an array, because OpenAPI
doesn't support lists of types.